+++ game/player.py
@@ -83,8 +83,8 @@ class Player(Person):
 
 
     def draw_need_camera(self,camera:Camera):
-        will_x = int(self.x) - int(camera.look_x) + console.get_max_width() / 2;
-        will_y = int(self.y) - int(camera.look_y) + console.get_max_height() / 2;
+        will_x = int(self.x) - int(camera.look_x) + int(console.get_max_width() / 2);
+        will_y = int(self.y) - int(camera.look_y) + int(console.get_max_height() / 2);
 
         console.move_cursor(int(will_x),int(will_y));
         print(self._style);
@@ -99,8 +99,8 @@ class Player(Person):
 
 
     def draw_in_cache_need_camera(self,camera:Camera):
-        will_x = int(self.x) - int(camera.look_x) + console.get_max_width() / 2;
-        will_y = int(self.y) - int(camera.look_y) + console.get_max_height() / 2;
+        will_x = int(self.x) - int(camera.look_x) + int(console.get_max_width() / 2);
+        will_y = int(self.y) - int(camera.look_y) + int(console.get_max_height() / 2);
 
 
         console.move_cache_cursor(int(will_x),int(will_y));
